Moved plugin centric plugins to buildsrc
