+++ buildSrc/src/main/kotlin/org/openmicroscopy/artifact/PluginProjectPlugin.kt
@@ -3,11 +3,13 @@ package org.openmicroscopy.artifact
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.*
-import org.gradle.plugin.devel.plugins.JavaGradlePluginPlugin
 
 class PluginProjectPlugin : Plugin<Project> {
     override fun apply(project: Project): Unit = project.run {
-        apply<JavaGradlePluginPlugin>()
-        apply<ArtifactPlugin>()
+        apply<AdditionalArtifactsPlugin>()
+        apply<AdditionalRepositoriesPlugin>()
+        apply<PluginProjectRepositoriesPlugin>()
+        apply<PluginPublishingPlugin>()
+        apply<FunctionalTestPlugin>()
     }
 }
