+++ buildSrc/src/main/kotlin/org/openmicroscopy/artifact/PluginProjectRepositoriesPlugin.kt
@@ -1,0 +1,17 @@
+package org.openmicroscopy.artifact
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.repositories
+import java.net.URI
+
+class PluginProjectRepositoriesPlugin : Plugin<Project> {
+    override fun apply(project: Project): Unit = project.run {
+        repositories {
+            maven {
+                name = "ome.maven"
+                url = URI.create("https://artifacts.openmicroscopy.org/artifactory/maven/")
+            }
+        }
+    }
+}
