+++ src/main/kotlin/org/openmicroscopy/artifact/ProjectPlugin.kt
@@ -23,11 +23,13 @@ package org.openmicroscopy.artifact
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.apply
 
 class ProjectPlugin : Plugin<Project> {
     override fun apply(project: Project): Unit = project.run {
-        apply<ArtifactPlugin>()
-        apply<ProjectRepositoriesPlugin>()
+        apply<AdditionalArtifactsPlugin>()
+        apply<AdditionalRepositoriesPlugin>()
+        apply<PluginPublishingPlugin>()
+        apply<PublishingPlugin>()
     }
 }
