+++ src/main/kotlin/org/openmicroscopy/artifact/PublishingPlugin.kt
